changes made to function init
This function is comprised of two click events which is why I need to load all letter elements and their values before any event occurred.  I was running into problems where the letter buttons didn't work because loadLetters was previously called in my first click event. The problem was that my second click event required that all letters be loaded BEFORE the first click event. At least I'm pretty sure that's what was wrong. This might be a little hard to understand, but this is my first time with more than one click event listener in a function, and it presented some interesting challenges.