changes made to letter button click event listener in init
Now, if a letter is correct, the letter box gets highlighted yellow and updateGuessBoard is called. If a letter is incorrect then state.wrongGuessCount is iterated and the next part of the hangman is drawn (drawHangman is called).

Now with each click of a letter box, winCheck is called. The condition on line 288 could be changed so that winCheck is called less often, but I'm assuming the performance hit is neglibible.