Fixed having to splice through state.remainingLetters twice
I followed the advice given and now all letters are loaded into the "letter-selector" element with one arrow function instead of two. This had a cascading effect on the function "clearLetters" which now only needed to take one argument instead of two for each row. To compensate for styling, flex-wrap: wrap as been applied to the "letter-selector" class as well as left and right margins of 25% to force the letters to wrap somewhat. I was also able to clean up some of the HTML this way.