added a play again button
This button makes itself visible when the "playAgain" function is called. Once clicked, the function "gameReset" is called which makes the difficulty buttons visible, clears the guess box of all contents, resets the guess counters in state, and resets the hangman. After "gameReset", init is called and the game repeats itself.